{{ cookiecutter.repo_name }}/requirements.txt
coverage
detect-secrets==1.0.3
myst-parser
pre-commit
pytest
pytest-mock
python-dotenv
Sphinx
